+++ filters/src/main/java/com/matecat/converter/Main.java
@@ -1,5 +1,6 @@
 package com.matecat.converter;
 
+import ch.qos.logback.classic.LoggerContext;
 import com.matecat.converter.server.MatecatConverterServer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,19 +14,22 @@ public class Main {
 
 
 	public static void main(String[] args) throws Exception {
-		if (Charset.defaultCharset() != StandardCharsets.UTF_8) {
-			throw new Exception("Java default charset is " + Charset.defaultCharset() + ", must be UTF-8. Fix your configuration.");
-		}
+        if (Charset.defaultCharset() != StandardCharsets.UTF_8) {
+            throw new Exception("Java default charset is " + Charset.defaultCharset() + ", must be UTF-8. Fix your configuration.");
+        }
 
-		// Init the server
-		MatecatConverterServer server = new MatecatConverterServer();
+        // Init the server
+        MatecatConverterServer server = new MatecatConverterServer();
 
-		// Shutdown gracefully when receiving SIGTERM or similar
-		Runtime.getRuntime().addShutdownHook(new Thread(() -> {
-			LOGGER.info("Shutdown signal received, stopping the server...");
-			server.stop();
-			LOGGER.info("Server stopped successfully. Good bye!");
-		}));
+        // Shutdown gracefully when receiving SIGTERM or similar
+        Runtime.getRuntime().addShutdownHook(new Thread(() -> {
+            LOGGER.info("Shutdown signal received, stopping the server...");
+            server.stop();
+            LOGGER.info("Server stopped successfully. Good bye!");
+            // clean stopping of the logging library
+            LoggerContext loggerContext = (LoggerContext) LoggerFactory.getILoggerFactory();
+            loggerContext.stop();
+        }));
 	}
 
 }